Fixing graphics lib compile warnings

+++ libs/graphics/jswrap_graphics.h
@@ -118,7 +118,7 @@ void _jswrap_graphics_freeImageInfo(GfxDrawImageInfo *info);
 /** Parse an image into GfxDrawImageInfo. See drawImage for image format docs. Returns true on success.
  * if 'image' is a string or ArrayBuffer, imageOffset is the offset within that (usually 0)
  */
-bool _jswrap_graphics_parseImage(JsGraphics *gfx, JsVar *image, unsigned int imageOffset, GfxDrawImageInfo *info);
+bool _jswrap_graphics_parseImage(JsGraphics *gfx, JsVar *image, size_t imageOffset, GfxDrawImageInfo *info);
 
 
 /// This is for rotating and scaling layers
